+++ src/ResourceManager/Common/Commands.ScenarioTests.ResourceManager.Common/RMTestBase.cs
@@ -13,18 +13,15 @@
 // ----------------------------------------------------------------------------------
 
 using Microsoft.Azure.Commands.Common.Authentication;
+using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Commands.Common.Authentication.Models;
-using Microsoft.WindowsAzure.Commands.Common;
+using Microsoft.Azure.Commands.ScenarioTest;
 using Microsoft.WindowsAzure.Commands.Common.Test.Mocks;
 using Microsoft.WindowsAzure.Commands.Utilities.Common;
+using Moq;
 using System;
-using System.Collections.Generic;
 using System.Management.Automation;
 using System.Threading;
-using Moq;
-using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
-using Microsoft.Azure.Commands.ResourceManager.Common;
-using Microsoft.Azure.Commands.ScenarioTest;
 
 namespace Microsoft.WindowsAzure.Commands.Test.Utilities.Common
 {